# GH-4001 Optimize EffectiveTarget