@hmottestad 2022-06-21
## Before
```
Benchmark                                    (NUMBER_OF_TRANSACTIONS)  (existingTargets)  Mode  Cnt     Score   Error  Units
ClassBenchmarkEmpty.shacl                                         N/A                N/A  avgt    2    29.730          ms/op
ClassBenchmarkEmpty.shaclBulk                                     N/A                N/A  avgt    2    22.942          ms/op
ComplexTargetBenchmark.benchmark                                  N/A                  1  avgt    2    24.155          ms/op
ComplexTargetBenchmark.benchmark                                  N/A               1000  avgt    2    23.071          ms/op
ComplexTargetBenchmark.benchmark                                  N/A             100000  avgt    2    21.143          ms/op
DatatypeBenchmarkLinear.shacl                                       1                N/A  avgt    2     0.547          ms/op
DatatypeBenchmarkLinear.shacl                                      10                N/A  avgt    2     4.071          ms/op
DatatypeBenchmarkLinear.shacl                                     100                N/A  avgt    2    35.581          ms/op
DatatypeBenchmarkPrefilled.shacl                                  N/A                N/A  avgt    2    20.650          ms/op
HasValueBenchmarkEmpty.shacl                                      N/A                N/A  avgt    2    51.198          ms/op
HasValueBenchmarkEmpty.shaclBulk                                  N/A                N/A  avgt    2    31.474          ms/op
MinCountBenchmarkEmpty.shacl                                      N/A                N/A  avgt    2    17.835          ms/op
MinCountBenchmarkEmpty.shaclBulk                                  N/A                N/A  avgt    2    11.771          ms/op
MinCountBenchmarkEmpty.shaclClear                                 N/A                N/A  avgt    2    33.562          ms/op
MinCountBenchmarkEmpty.shaclMinCountZero                          N/A                N/A  avgt    2     8.523          ms/op
MinCountBenchmarkPrefilled.shacl                                  N/A                N/A  avgt    2    23.809          ms/op
NotClassBenchmarkEmpty.shacl                                      N/A                N/A  avgt    2   121.219          ms/op
NotClassBenchmarkEmpty.shaclBulk                                  N/A                N/A  avgt    2    55.387          ms/op
NotClassBenchmarkEmpty.shaclWithoutAnimals                        N/A                N/A  avgt    2    71.523          ms/op
NotMaxCountBenchmarkEmpty.shacl                                   N/A                N/A  avgt    2    58.103          ms/op
NotMaxCountBenchmarkEmpty.shaclBulk                               N/A                N/A  avgt    2    16.474          ms/op
OrDatatypeBenchmark.shacl                                         N/A                N/A  avgt    2   197.811          ms/op
OrDatatypeBenchmark.shaclBulk                                     N/A                N/A  avgt    2    57.463          ms/op
QualifiedValueShapeBenchmarkEmpty.shacl                           N/A                N/A  avgt    2  4937.130          ms/op
QualifiedValueShapeBenchmarkEmpty.shaclBulk                       N/A                N/A  avgt    2   257.220          ms/op
RdfsReasonerBenchmarkEmpty.shacl                                  N/A                N/A  avgt    2   107.487          ms/op
RdfsReasonerBenchmarkEmpty.shaclBulk                              N/A                N/A  avgt    2    41.374          ms/op
TargetShapeBenchmark.shacl                                        N/A                N/A  avgt    2    29.831          ms/op
ValueInBenchmarkEmpty.shacl                                       N/A                N/A  avgt    2    52.754          ms/op
ValueInBenchmarkEmpty.shaclBulk                                   N/A                N/A  avgt    2    41.566          ms/op
```


## After
```
Benchmark                                    (NUMBER_OF_TRANSACTIONS)  (existingTargets)  Mode  Cnt     Score   Error  Units
ClassBenchmarkEmpty.shacl                                         N/A                N/A  avgt    2    30.633          ms/op
ClassBenchmarkEmpty.shaclBulk                                     N/A                N/A  avgt    2    23.485          ms/op
ComplexTargetBenchmark.benchmark                                  N/A                  1  avgt    2    22.442          ms/op
ComplexTargetBenchmark.benchmark                                  N/A               1000  avgt    2    22.302          ms/op
ComplexTargetBenchmark.benchmark                                  N/A             100000  avgt    2    21.044          ms/op
DatatypeBenchmarkLinear.shacl                                       1                N/A  avgt    2     0.564          ms/op
DatatypeBenchmarkLinear.shacl                                      10                N/A  avgt    2     4.365          ms/op
DatatypeBenchmarkLinear.shacl                                     100                N/A  avgt    2    37.680          ms/op
DatatypeBenchmarkPrefilled.shacl                                  N/A                N/A  avgt    2    20.933          ms/op
HasValueBenchmarkEmpty.shacl                                      N/A                N/A  avgt    2    46.016          ms/op
HasValueBenchmarkEmpty.shaclBulk                                  N/A                N/A  avgt    2    31.838          ms/op
MinCountBenchmarkEmpty.shacl                                      N/A                N/A  avgt    2    10.833          ms/op
MinCountBenchmarkEmpty.shaclBulk                                  N/A                N/A  avgt    2    12.083          ms/op
MinCountBenchmarkEmpty.shaclClear                                 N/A                N/A  avgt    2    30.959          ms/op
MinCountBenchmarkEmpty.shaclMinCountZero                          N/A                N/A  avgt    2     8.548          ms/op
MinCountBenchmarkPrefilled.shacl                                  N/A                N/A  avgt    2    14.247          ms/op
NotClassBenchmarkEmpty.shacl                                      N/A                N/A  avgt    2   113.696          ms/op
NotClassBenchmarkEmpty.shaclBulk                                  N/A                N/A  avgt    2    55.073          ms/op
NotClassBenchmarkEmpty.shaclWithoutAnimals                        N/A                N/A  avgt    2    70.212          ms/op
NotMaxCountBenchmarkEmpty.shacl                                   N/A                N/A  avgt    2    50.259          ms/op
NotMaxCountBenchmarkEmpty.shaclBulk                               N/A                N/A  avgt    2    16.788          ms/op
OrDatatypeBenchmark.shacl                                         N/A                N/A  avgt    2   198.668          ms/op
OrDatatypeBenchmark.shaclBulk                                     N/A                N/A  avgt    2    59.711          ms/op
QualifiedValueShapeBenchmarkEmpty.shacl                           N/A                N/A  avgt    2  3735.261          ms/op
QualifiedValueShapeBenchmarkEmpty.shaclBulk                       N/A                N/A  avgt    2   250.343          ms/op
RdfsReasonerBenchmarkEmpty.shacl                                  N/A                N/A  avgt    2   105.352          ms/op
RdfsReasonerBenchmarkEmpty.shaclBulk                              N/A                N/A  avgt    2    42.201          ms/op
TargetShapeBenchmark.shacl                                        N/A                N/A  avgt    2    29.755          ms/op
ValueInBenchmarkEmpty.shacl                                       N/A                N/A  avgt    2    43.488          ms/op
ValueInBenchmarkEmpty.shaclBulk                                   N/A                N/A  avgt    2    41.616          ms/op
```